requirements.txt
djangorestframework-simplejwt==4.6.0
drf-yasg2==1.19.4
factory_boy==3.2.0
fakeredis==1.6.1
Faker==3.0.1
mock==4.0.2
phonenumbers==8.13.42
pillow==10.4.0
pycountry==24.6.1
psycopg2-binary==2.9.9
pydantic==2.8.2
PyJWT==2.9.0
redis==3.5.3
requests==2.32.3
six==1.16.0
celery==5.4.0
redis==3.5.3
Django==3.1.2
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-cities==0.6.2
django-model-utils==4.2.0
django-phonenumber-field==5.2.0
django-timezone-field==4.2.3
djangorestframework==3.12.1
gunicorn==20.0.4
django-cors-headers==3.6.0
kombu==5.4.0
pypdf2==3.0.1
reportlab==4.0.9